+++ docs/changelog.md
@@ -89,6 +89,59 @@
 
 ## 更新说明详情
 
+1.18.20 fix: fix stock_individual_spot_xq interface
+
+	1. 修复 stock_individual_spot_xq 接口
+
+1.18.19 fix: fix GitHub action
+
+	1. 修复配置文件
+
+1.18.18 fix: fix stock_individual_spot_xq interface
+
+	1. 修复 stock_individual_spot_xq 接口
+
+1.18.17 fix: fix stock_individual_spot_xq interface
+
+	1. 修复 stock_individual_spot_xq 接口
+
+1.18.16 fix: fix news_economic_baidu interface
+
+	1. 修复 news_economic_baidu 接口
+
+1.18.15 fix: fix fund_etf_scale_sse and fund_etf_scale_szse interface
+
+	1. 修复 fund_etf_scale_sse 接口-基金份额字段数据解析错误，统一使用"份"作为单位
+	2. 修复 fund_etf_scale_szse 接口-基金份额字段数据解析错误，统一使用"份"作为单位
+
+1.18.14 add: add fund_etf_scale_sse and fund_etf_scale_szse interface
+
+	1. 新增 fund_etf_scale_sse 接口-上海证券交易所 ETF 基金份额数据
+	2. 新增 fund_etf_scale_szse 接口-深圳证券交易所 ETF 基金份额数据
+
+1.18.13 fix: fix futures_spot_price_daily interface
+
+	1. 修复 futures_spot_price_daily 接口
+
+1.18.12 fix: fix news_economic_baidu interface
+
+	1. 修复 news_economic_baidu 接口
+
+1.18.11 fix: fix stock_individual_spot_xq interface
+
+	1. 修复 stock_individual_spot_xq 接口
+	2. 修复 stock_margin_ratio_pa 接口
+
+1.18.10 fix: fix futures_comm_info interface
+
+	1. 修复 futures_comm_info 接口
+	2. 修复 index_detail_hist_cni 接口
+
+1.18.9 fix: fix fx_quote_baidu interface
+
+	1. 修复 fx_quote_baidu 接口
+	2. 修复 stock_zh_a_daily 接口
+
 1.18.8 fix: fix stock_zh_valuation_comparison_em interface
 
 	1. 修复 stock_zh_valuation_comparison_em 接口
@@ -6056,6 +6109,30 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.18.20 fix: fix stock_individual_spot_xq interface
+
+1.18.19 fix: fix GitHub action
+
+1.18.18 fix: fix stock_individual_spot_xq interface
+
+1.18.17 fix: fix stock_individual_spot_xq interface
+
+1.18.16 fix: fix news_economic_baidu interface
+
+1.18.15 fix: fix fund_etf_scale_sse and fund_etf_scale_szse interface
+
+1.18.14 add: add fund_etf_scale_sse and fund_etf_scale_szse interface
+
+1.18.13 fix: fix futures_spot_price_daily interface
+
+1.18.12 fix: fix news_economic_baidu interface
+
+1.18.11 fix: fix stock_individual_spot_xq interface
+
+1.18.10 fix: fix futures_comm_info interface
+
+1.18.9 fix: fix fx_quote_baidu interface
 
 1.18.8 fix: fix stock_zh_valuation_comparison_em interface